requirements.txt
findspark==2.0.1
python-dotenv==0.20.0
